--- lib/features/home/logic/connectivity_cubit/connectivity_state.dart
@@ -1,10 +1,0 @@
-part of 'connectivity_cubit.dart';
-
-sealed class ConnectivityState extends Equatable {
-  const ConnectivityState();
-}
-
-final class ConnectivityInitial extends ConnectivityState {
-  @override
-  List<Object> get props => [];
-}
