+++ lib/features/home/ui/widgets/currency_exchange_component/connectivity_status_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import '../../../core/utils/app_colors.dart';
-import '../../../core/utils/app_text_style.dart';
+import '../../../../../core/utils/app_colors.dart';
+import '../../../../../core/utils/app_text_style.dart';
 
 class ConnectivityStatusWidget extends StatelessWidget {
   const ConnectivityStatusWidget({
